+++ swing/framework-ui-tools/src/test/java/is/codion/swing/framework/ui/tools/generator/DomainGeneratorPanelTest.java
@@ -1,24 +1,24 @@
 /*
  * Copyright (c) 2012 - 2024, Björn Darri Sigurðsson. All Rights Reserved.
  */
-package is.codion.swing.framework.ui.tools.explorer;
+package is.codion.swing.framework.ui.tools.generator;
 
 import is.codion.common.db.database.Database;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.user.User;
-import is.codion.swing.framework.model.tools.explorer.DatabaseExplorerModel;
+import is.codion.swing.framework.model.tools.generator.DomainGeneratorModel;
 
 import org.junit.jupiter.api.Test;
 
-public class DatabaseExplorerPanelTest {
+public class DomainGeneratorPanelTest {
 
   private static final User UNIT_TEST_USER =
           User.parse(System.getProperty("codion.test.user", "scott:tiger"));
 
   @Test
   void test() throws DatabaseException {
-    DatabaseExplorerModel model = DatabaseExplorerModel.databaseExplorerModel(Database.instance(), UNIT_TEST_USER);
-    new DatabaseExplorerPanel(model);
+    DomainGeneratorModel model = DomainGeneratorModel.domainGeneratorModel(Database.instance(), UNIT_TEST_USER);
+    new DomainGeneratorPanel(model);
     model.schemaModel().refresh();
     model.schemaModel().sortItems();
     model.schemaModel().selectionModel().setSelectedIndex(2);